Rename ExternalProject -> ProjectReference
